+++ src/frontend/js/utils/api/lms/base.spec.ts
@@ -1,20 +1,56 @@
-import { ContextFactory } from 'utils/test/factories';
-import { APIBackend } from 'types/api';
+import { ContextFactory as mockContextFactory } from 'utils/test/factories';
+import context from 'utils/context';
+import { location } from 'utils/indirection/window';
+import API from './base';
 
-describe('Base API', () => {
-  const context = ContextFactory({
+jest.mock('utils/indirection/window', () => ({
+  location: {
+    reload: jest.fn(),
+  },
+}));
+
+jest.mock('utils/context', () => ({
+  __esModule: true,
+  default: mockContextFactory({
+    authentication: {
+      backend: 'base',
+      endpoint: 'https://demo.endpoint/api',
+    },
     lms_backends: [
       {
-        backend: APIBackend.BASE,
+        backend: 'base',
         course_regexp: '(?<course_id>.*)',
         endpoint: 'https://demo.endpoint/api',
       },
     ],
-  }).generate();
-  window.__richie_frontend_context__ = { context };
-  const { default: API } = require('./base');
-  const LMSConf = context.lms_backends[0];
+  }).generate(),
+}));
+
+describe('Base API', () => {
+  const LMSConf = context.lms_backends![0];
   const BaseAPI = API(LMSConf);
+
+  beforeEach(() => {
+    jest.resetAllMocks();
+  });
+
+  describe('user', () => {
+    it('simulates that authenticated user is admin', async () => {
+      const response = await BaseAPI.user.me();
+      expect(response?.username).toBe('admin');
+      expect(response?.access_token).toBeDefined();
+    });
+
+    it('reloads the page when login is called', async () => {
+      BaseAPI.user.login();
+      expect(location.reload).toHaveBeenCalledTimes(1);
+    });
+
+    it('reloads the page when register is called', async () => {
+      BaseAPI.user.register();
+      expect(location.reload).toHaveBeenCalledTimes(1);
+    });
+  });
 
   describe('enrollment', () => {
     beforeEach(() => {
@@ -34,7 +70,7 @@ describe('Base API', () => {
           username: 'johndoe',
         });
 
-        const response = await BaseAPI.enrollment.get(
+        const response: any = await BaseAPI.enrollment.get(
           'https://demo.endpoint/courses?course_id=af1987efz98:afe78',
           { username: 'johndoe' },
         );
@@ -72,8 +108,9 @@ describe('Base API', () => {
       });
 
       it('returns false if user is anonymous', async () => {
-        const enrollment = await BaseAPI.enrollment.get(
+        const enrollment: any = await BaseAPI.enrollment.get(
           'https://demo.endpoint/courses?course_id=af1987efz98:afe78',
+          { username: 'johndoe' },
         );
 
         const response = await BaseAPI.enrollment.isEnrolled(enrollment);